+++ module/helpers/effects.mjs
@@ -1,3 +1,4 @@
+import SPACE1889Helper from "./helper.mjs";
 /**
  * Manage Active Effect instances through the Actor Sheet via effect control buttons.
  * @param {MouseEvent} event      The left-click event on the effect control
@@ -12,18 +13,30 @@ export function onManageActiveEffect(event, owner) {
 		case "create":
 			const gameRound = game.combat ? game.combat.round : 0;
 			const gameTurn = game.combat ? game.combat.turn : 0;
-
-			return owner.createEmbeddedDocuments("ActiveEffect", [{
-				label: game.i18n.localize("SPACE1889.EffectNew"),
-				icon: "icons/svg/aura.svg",
-				origin: owner.uuid,
-				"duration.rounds": 1,
-				"duration.seconds": 6,
-				"duration.startRound": gameRound,
-				"duration.startTurn": gameTurn,
-				"duration.startTime": game.time.worldTime,
-				disabled: li.dataset.effectType === "inactive"
-			}]);
+			if (SPACE1889Helper.isFoundryV10Running())
+				return owner.createEmbeddedDocuments("ActiveEffect", [{
+					label: game.i18n.localize("SPACE1889.EffectNew"),
+					icon: "icons/svg/aura.svg",
+					origin: owner.uuid,
+					"duration.rounds": 1,
+					"duration.seconds": 6,
+					"duration.startRound": gameRound,
+					"duration.startTurn": gameTurn,
+					"duration.startTime": game.time.worldTime,
+					disabled: li.dataset.effectType === "inactive"
+				}]);
+			else
+				return owner.createEmbeddedDocuments("ActiveEffect", [{
+					name: game.i18n.localize("SPACE1889.EffectNew"),
+					icon: "icons/svg/aura.svg",
+					origin: owner.uuid,
+					"duration.rounds": 1,
+					"duration.seconds": 6,
+					"duration.startRound": gameRound,
+					"duration.startTurn": gameTurn,
+					"duration.startTime": game.time.worldTime,
+					disabled: li.dataset.effectType === "inactive"
+				}]);
 		case "edit":
 			return effect.sheet.render(true);
 		case "delete":
@@ -33,38 +46,3 @@ export function onManageActiveEffect(event, owner) {
 	}
 }
 
-/**
- * Prepare the data structure for Active Effects which are currently applied to an Actor or Item.
- * @param {ActiveEffect[]} effects    The array of Active Effect instances to prepare sheet data for
- * @return {object}                   Data for rendering
- */
-export function prepareActiveEffectCategories(effects) {
-
-	// Define effect header categories
-	const categories = {
-		temporary: {
-			type: "temporary",
-			label: "Temporary Effects",
-			effects: []
-		},
-		passive: {
-			type: "passive",
-			label: "Passive Effects",
-			effects: []
-		},
-		inactive: {
-			type: "inactive",
-			label: "Inactive Effects",
-			effects: []
-		}
-	};
-
-	// Iterate over active effects, classifying them into categories
-	for ( let e of effects ) {
-		e._getSourceName(); // Trigger a lookup for the source name
-		if ( e.disabled ) categories.inactive.effects.push(e);
-		else if ( e.isTemporary ) categories.temporary.effects.push(e);
-		else categories.passive.effects.push(e);
-	}
-	return categories;
-}