+++ module/space1889.mjs
@@ -313,6 +313,11 @@ Handlebars.registerHelper('isGm', function (str)
 	return game.user.isGM;
 })
 
+Handlebars.registerHelper('isFvttV10', function (str)
+{
+	return SPACE1889Helper.isFoundryV10Running();
+})
+
 /* -------------------------------------------- */
 /*  Ready Hook                                  */
 /* -------------------------------------------- */